Fix submit plan from own leader
