+++ src/pages/sykmeldt/[oppfolgingsdialogId]/seplanen.tsx
@@ -2,17 +2,18 @@ import { NextPage } from "next";
 import { useAktivPlanSM } from "../../../api/queries/sykmeldt/oppfolgingsplanerQueriesSM";
 import { SendTilGodkjenningSM } from "../../../components/seplanen/sendtilgodkjenning/SendTilGodkjenningSM";
 import { OppfolgingsplanOversikt } from "../../../components/seplanen/OppfolgingsplanOversikt";
-import { Alert } from "@navikt/ds-react";
 import { beskyttetSideUtenProps } from "../../../auth/beskyttetSide";
 import {
   OppfolgingsplanPageSM,
   Page,
 } from "../../../components/blocks/wrappers/oppfolgingsplanpagesm/OppfolgingsplanPageSM";
+import { GodkjennEgenPlan } from "../../../components/seplanen/sendtilgodkjenning/GodkjennEgenPlan";
+import React from "react";
 
 const Seplanen: NextPage = () => {
   const aktivPlan = useAktivPlanSM();
   const arbeidstakerFnr = aktivPlan?.arbeidstaker.fnr;
-  const narmesteLederFnr = aktivPlan?.arbeidsgiver?.naermesteLeder?.navn;
+  const narmesteLederFnr = aktivPlan?.arbeidsgiver?.naermesteLeder?.fnr;
 
   const isOwnLeader =
     narmesteLederFnr && arbeidstakerFnr && narmesteLederFnr === arbeidstakerFnr;
@@ -23,11 +24,8 @@ const Seplanen: NextPage = () => {
       {!isOwnLeader && aktivPlan && (
         <SendTilGodkjenningSM oppfolgingsplan={aktivPlan} />
       )}
-      {isOwnLeader && (
-        <Alert variant={"info"}>
-          Fordi du er din egen leder, må du logge inn som arbeidsgiver for å
-          fullføre planen.
-        </Alert>
+      {isOwnLeader && aktivPlan && (
+        <GodkjennEgenPlan oppfolgingsplan={aktivPlan} />
       )}
     </OppfolgingsplanPageSM>
   );
